cleared the storybook and material ui
