--- src/components/Login/LoginForm.stories.js
@@ -1,47 +1,0 @@
-import React from 'react'
-
-import { initialStateData, LoginForm } from './LoginForm'
-import { withKnobs, object } from '@storybook/addon-knobs/react'
-
-
-// attaches component to the story book
-export default {
-    component: LoginForm,
-    title: 'LoginForm',
-    decorators: [withKnobs], // use to display state in storybook
-    excludeStories: /.*Data$/, // export ending with Data wont be treated as a story
-}
-
-
-// states for stories
-const beforeRequestStateData = {
-    ...initialStateData,
-    email: 'crash.test.dummy@gmail.com',
-    password: 'crash.test',
-}
-
-const errorStateData = {
-    ...initialStateData,
-    error: {
-        "email": "This field is required.",
-        "password": "This field is required."
-    },
-}
-
-const loadingStateData = {
-    ...beforeRequestStateData,
-    loading: true
-}
-
-const edgeCaseErrStateData = {
-    ...beforeRequestStateData,
-    loading: false,
-    edgeCaseErr: 'Credentials did not match'
-}
-
-// creating the stories
-export const initialState = () => <LoginForm stateData={object('state', initialStateData)} />
-export const errorState = () => <LoginForm stateData={object('state', errorStateData)} />
-export const loadingState = () => <LoginForm stateData={object('state', loadingStateData)} />
-export const beforeRequestState = () => <LoginForm stateData={object('state', beforeRequestStateData)} />
-export const edgeCaseErrState = () => <LoginForm stateData={object('state', edgeCaseErrStateData)} />
