+++ src/components/Login/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import LoginForm from './LoginForm'
+
+export default function index() {
+  return (
+    <div>
+      <LoginForm />
+    </div>
+  )
+}
