--- src/components/Tasks/performancePreview.js
@@ -1,199 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import { Grid, Card, Typography, Divider, Box } from '@material-ui/core';
-
-//Error Component
-import Error from '../Widgets/Error/Error';
-//Loading
-import { Skeleton } from '@material-ui/lab';
-
-//Icons
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import GroupIcon from '@material-ui/icons/Group';
-import EqualizerIcon from '@material-ui/icons/Equalizer';
-
-const styles = {
-    root: {
-        margin: '50px auto',
-        textAlign: 'center',
-    },
-
-    icon_box: {
-        position: 'absolute',
-        margin: '-40px 0',
-        zIndex: +1,
-        height: '60px',
-        width: '60px',
-        border: 0,
-        borderRadius: '5px 5px 5px 5px',
-    },
-
-    cards: {
-        border: 0,
-        borderRadius: 5,
-        height: '100px',
-        width: '200px',
-        padding: 10,
-    },
-    icons: {
-        margin: '20px auto',
-        color: 'white',
-        fontSize: 20,
-    },
-};
-
-const performancePreview = ({
-    stateData: {
-        tasksCompleted,
-        totalTasks,
-        points,
-        totalPoints,
-        rank,
-        totalVolunteers,
-        error,
-        loading,
-    },
-    classes,
-}) => {
-    //If an error occurs and the page is not yet loading
-    if (error && !loading) {
-        const stateData = {
-            error:
-                'Something Went Wrong Please try again later or try refreshing',
-            width: '80%',
-            direction: 'column',
-            iconSize: 40,
-        };
-        return <Error stateData={stateData} />;
-    }
-    //Else return the normal state accordingly
-    return (
-        <div>
-            <Typography align='center' variant='h5'>
-                Performance
-            </Typography>
-
-            {!loading ? (
-                <Grid
-                    container
-                    spacing={5}
-                    direction='row'
-                    justify='center'
-                    alignItems='center'
-                    className={classes.root}>
-                    {/* Tasks */}
-                    <Grid item>
-                        <Card className={classes.cards} boxShadow={1}>
-                            <Box
-                                className={classes.icon_box}
-                                bgcolor='success.main'
-                                alignItems='center'
-                                boxShadow={1}>
-                                <AssignmentIcon className={classes.icons} />
-                            </Box>
-
-                            <Typography align='right' variant='h6'>
-                                <strong>Tasks</strong>
-                            </Typography>
-                            <Typography align='right' variant='h4'>
-                                {tasksCompleted}
-                            </Typography>
-
-                            <Divider light />
-
-                            <Typography align='left' variant='body1'>
-                                <Box p={1}>Total Tasks: {totalTasks}</Box>
-                            </Typography>
-                        </Card>
-                    </Grid>
-                    {/* Points */}
-                    <Grid item>
-                        <Card className={classes.cards} boxShadow={1}>
-                            <Box
-                                className={classes.icon_box}
-                                bgcolor='secondary.main'
-                                boxShadow={1}>
-                                <EqualizerIcon className={classes.icons} />
-                            </Box>
-
-                            <Typography align='right' variant='h6'>
-                                <strong>Points</strong>
-                            </Typography>
-                            <Typography align='right' variant='h4'>
-                                {points}
-                            </Typography>
-
-                            <Divider light />
-
-                            <Typography align='left' variant='body1'>
-                                <Box p={1}>Max points: {totalPoints}</Box>
-                            </Typography>
-                        </Card>
-                    </Grid>
-                    {/* Leader Board */}
-                    <Grid item>
-                        <Card className={classes.cards} boxShadow={1}>
-                            <Box
-                                className={classes.icon_box}
-                                bgcolor='info.main'
-                                boxShadow={1}>
-                                <GroupIcon className={classes.icons} />
-                            </Box>
-
-                            <Typography align='right' variant='h6'>
-                                <strong>Rank</strong>
-                            </Typography>
-                            <Typography align='right' variant='h4'>
-                                {rank}
-                            </Typography>
-
-                            <Divider light />
-
-                            <Typography align='left' variant='body1'>
-                                <Box p={1}>
-                                    Total Volunteers: {totalVolunteers}
-                                </Box>
-                            </Typography>
-                        </Card>
-                    </Grid>
-                </Grid>
-            ) : (
-                //Loading State
-                <Grid
-                    container
-                    spacing={5}
-                    direction='row'
-                    justify='center'
-                    alignItems='center'
-                    className={classes.root}>
-                    <Grid item>
-                        <Skeleton variant='rect' width={200} height={100} />
-                    </Grid>
-                    <Grid item>
-                        <Skeleton variant='rect' width={200} height={100} />
-                    </Grid>
-                    <Grid item>
-                        <Skeleton variant='rect' width={200} height={100} />
-                    </Grid>
-                </Grid>
-            )}
-        </div>
-    );
-};
-
-performancePreview.propTypes = {
-    stateData: PropTypes.shape({
-        tasksCompleted: PropTypes.number.isRequired,
-        totalTasks: PropTypes.number.isRequired,
-        points: PropTypes.number.isRequired,
-        totalPoints: PropTypes.number.isRequired,
-        rank: PropTypes.number.isRequired,
-        totalVolunteers: PropTypes.number.isRequired,
-        loading: PropTypes.bool.isRequired,
-        error: PropTypes.bool.isRequired,
-    }),
-    classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(performancePreview);
