+++ src/utils/CustomHistory.js
@@ -1,3 +1,5 @@
 import { createBrowserHistory } from 'history'
 
-export default createBrowserHistory()
+const history = createBrowserHistory();
+
+export default history;
