+++ src/utils/RootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import LoginReducer from '../actions/LoginActions'
+import { LoginReducer } from '../components/Login/actions'
 
 
 export default combineReducers({
